new build & ranking poster

+++ build/Modules/DeleteAll.js
@@ -7,28 +7,40 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
         step((generator = generator.apply(thisArg, _arguments || [])).next());
     });
 };
-const deleteAll = function (client) {
-    return __awaiter(this, void 0, void 0, function* () {
-        let Globalcmds = yield client.api.applications(client.user.id).commands.get();
-        let Guildcmds = yield client.api
-            .applications(client.user.id)
-            .guilds(process.env.GUILD_ID)
-            .commands.get();
-        Globalcmds.forEach(cmd => {
-            client.api
-                .applications(client.user.id)
-                .commands(cmd.id)
-                .delete();
-            console.log("Deleted Global command { name: " + cmd.name + ", ID: " + cmd.id + " }");
+class deleteAll {
+    constructor(client) {
+        this.client = client;
+    }
+    run() {
+        return __awaiter(this, void 0, void 0, function* () {
+            let client = this.client;
+            if (client.deleteCmd) {
+                let Globalcmds = yield client.api.applications(client.user.id).commands.get();
+                let Guildcmds = yield client.api
+                    .applications(client.user.id)
+                    .guilds(process.env.GUILD_ID)
+                    .commands.get();
+                Globalcmds.forEach(cmd => {
+                    client.api
+                        .applications(client.user.id)
+                        .commands(cmd.id)
+                        .delete();
+                    console.log("Deleted Global command { name: " + cmd.name + ", ID: " + cmd.id + " }");
+                });
+                Guildcmds.forEach(cmds => {
+                    client.api
+                        .applications(client.user.id)
+                        .guilds(process.env.GUILD_ID)
+                        .commands(cmds.id)
+                        .delete();
+                    console.log("Deleted Guild command { name: " + cmds.name + ", ID: " + cmds.id + " }");
+                });
+                return 1;
+            }
+            else {
+                return 0;
+            }
         });
-        Guildcmds.forEach(cmds => {
-            client.api
-                .applications(client.user.id)
-                .guilds(process.env.GUILD_ID)
-                .commands(cmds.id)
-                .delete();
-            console.log("Deleted Guild command { name: " + cmds.name + ", ID: " + cmds.id + " }");
-        });
-    });
-};
+    }
+}
 module.exports = deleteAll;
